Comments and room.exr specificity
The piUnitsConvert needs to be upgraded.  I put in some more suggestions.  I am not even sure the function is properly thought through.  If it is needed, we should write some tests for it.

There are multiple copies of room.exr floating around.  So in the future we should be specific that we want the one that is in the data/lights directory.  Hopefully that file is properly copied to local and during the rsync.