A lot of cleaning for human eye case
I changed nearly all references to realisticEye to become humaneye.  Still a couple to go.

I starting dealing with accommodation as the proper special case for human eye model.

There are still some bugs on the ISET3d-v4 side, but it did render.

Zhenyi needs to make the spectralpath part work for chromatic aberration.  And we need testing.  And then we can integrate the TG retina surface shape.