Fixed a material delete
The use of containers for materials meant that removing a material needed an update.

Also, the white surface loads with an empty material.  I remove it in piRecipeCreate.  It is still there in piRecipeDefault.

There were stray old material scripts that I deleted.

I don't understand when spectrum tex1 and spectrum tex2 are used.  They seem not to be part of checkerboard texture, but maybe they are?  Zheng might be able to answer this.