Simplifying the ChessSet file organization
And trying to make it work with the use remote resources flag.  Which I added 'remote resources' as an equivalent.

When we simply copy this ChessSet directory over to local, it will render.  But something in the piWrite() function breaks that so we do not see the planeTan and planeBrown materials.