Managing names, simplifying code
We now have a situation where there is
AttributeBegin
  Stuff
  AttributeBegin
    More stuff
  AttributeEnd
AttributeEnd

And the name and transform is in the first Begin/End but not assigned nicely because it is not visible in the recursive call.  Maybe I can put it back using the 'name' argument (3rd) that I deleted in parseGeometryText.