Simplified parseGeometryText again
This time I removed the confusing initial block whose main purpose, I think, was to make sure that different instances of the same object had different names.

The way we rename may change, but for the moment I handle this at the end of piRead.

I am running through a lot of scenes now, to see if I broken anything.  Looking pretty good.