Added test on empty subtree return
Hard to understand why this ever worked, really. Possibly because we weren't properly testing for an identity transform in the returned branch node.

As things stand, we are adding in a lot of identity branch nodes.  They seem useless to me.  But we there are cases (maybe instances?) that require them.