Comments, setting eye model retina distance
The new script t_eyeRetinaDistance is used to figure how to set the retina distance for the different eye models.

At the moment, it is set by hand.  A far away object, and move the retina systematically until we get a chromatic aberration with mostly blue, good in the green, a little in the red.