Many comments and parameter setting
The retina distance parameter seems OK now.

The accommodation value we set for the eye models seems a bit off.  So when we set the human eye model to accommodate for some distance, it is really accommodated for a slightly different distance.

I am considering improving this.  See notes in the files.