Experimenting with area lights
We need to rotate the shape of the area so we can see more or less of it.  We need to figure out the direction it is pointing!  We need to figure out the point2 uv and normal and such.  We need to change the shape of the area light and control its size.

We also learned that when it is a rectangle, specscale increases the intensity by the square, not linearly.