Light keys and warning
The short and long mapLgt keys may not have the same number of elements.  This is to make certain operations with groups of lights easier for cars.  We are thinking about this.  This caused a problem with getting light names and ids, and that is now fixed.

Eliminated the warning about the identity geometry transform being skipped.